Move Cache interface one level up
